+++ keras_cv/models/backbones/csp_darknet/csp_darknet_backbone_presets_test.py
@@ -17,7 +17,12 @@ import pytest
 import tensorflow as tf
 from absl.testing import parameterized
 
-from keras_cv.models.backbones.csp_darknet import csp_darknet_backbone
+from keras_cv.models.backbones.csp_darknet.csp_darknet_aliases import (
+    CSPDarkNetMBackbone,
+)
+from keras_cv.models.backbones.csp_darknet.csp_darknet_backbone import (
+    CSPDarkNetBackbone,
+)
 
 
 @pytest.mark.large
@@ -32,15 +37,11 @@ class CSPDarkNetPresetSmokeTest(tf.test.TestCase, parameterized.TestCase):
         self.input_batch = tf.ones(shape=(2, 224, 224, 3))
 
     def test_backbone_output(self):
-        model = csp_darknet_backbone.CSPDarkNetBackbone.from_preset(
-            "csp_darknet_tiny"
-        )
+        model = CSPDarkNetBackbone.from_preset("csp_darknet_tiny")
         model(self.input_batch)
 
     def test_backbone_output_with_weights_tiny(self):
-        model = csp_darknet_backbone.CSPDarkNetBackbone.from_preset(
-            "csp_darknet_tiny_imagenet"
-        )
+        model = CSPDarkNetBackbone.from_preset("csp_darknet_tiny_imagenet")
         outputs = model(tf.ones(shape=(1, 512, 512, 3)))
 
         # The forward pass from a preset should be stable!
@@ -56,34 +57,30 @@ class CSPDarkNetPresetSmokeTest(tf.test.TestCase, parameterized.TestCase):
         )
 
     def test_applications_model_output(self):
-        model = csp_darknet_backbone.CSPDarkNetMBackbone()
+        model = CSPDarkNetMBackbone()
         model(self.input_batch)
 
     def test_applications_model_output_with_preset(self):
-        model = csp_darknet_backbone.CSPDarkNetBackbone.from_preset(
-            "csp_darknet_tiny_imagenet"
-        )
+        model = CSPDarkNetBackbone.from_preset("csp_darknet_tiny_imagenet")
         model(self.input_batch)
 
     def test_preset_docstring(self):
         """Check we did our docstring formatting correctly."""
-        for name in csp_darknet_backbone.CSPDarkNetBackbone.presets:
+        for name in CSPDarkNetBackbone.presets:
             self.assertRegex(
-                csp_darknet_backbone.CSPDarkNetBackbone.from_preset.__doc__,
+                CSPDarkNetBackbone.from_preset.__doc__,
                 name,
             )
 
     def test_unknown_preset_error(self):
         # Not a preset name
         with self.assertRaises(ValueError):
-            csp_darknet_backbone.CSPDarkNetBackbone.from_preset(
-                "unknown_weights"
-            )
+            CSPDarkNetBackbone.from_preset("unknown_weights")
 
     def test_load_weights_error(self):
         # Try to load weights when none available
         with self.assertRaises(ValueError):
-            csp_darknet_backbone.CSPDarkNetBackbone.from_preset(
+            CSPDarkNetBackbone.from_preset(
                 "csp_darknet_tiny", load_weights=True
             )
 
@@ -92,13 +89,13 @@ class CSPDarkNetPresetSmokeTest(tf.test.TestCase, parameterized.TestCase):
 class CSPDarkNetPresetFullTest(tf.test.TestCase, parameterized.TestCase):
     """
     Test the full enumeration of our preset.
-    This every presets for CSPDarkNet and is only run manually.
+    This tests every presets for CSPDarkNet and is only run manually.
     Run with:
     `pytest keras_cv/models/backbones/csp_darknet/csp_darknet_backbone_presets_test.py --run_extra_large`  # noqa: E501
     """
 
     def test_load_csp_darknet(self):
         input_data = tf.ones(shape=(2, 512, 512, 3))
-        for preset in csp_darknet_backbone.CSPDarkNetBackbone.presets:
-            model = csp_darknet_backbone.CSPDarkNetBackbone.from_preset(preset)
+        for preset in CSPDarkNetBackbone.presets:
+            model = CSPDarkNetBackbone.from_preset(preset)
             model(input_data)
